[Feat/AdminPage] 체크박스 컴포넌트 분리
